fix : Format error in AlgorithmData.ClearRate

+++ World Tower Defense/Assets/Scripts/Level/AlogrithmData.cs
@@ -17,7 +17,7 @@ namespace GameData
             //아래를 통해 클리어 시간을 산출 가능
             public float fitnessClearTimeRate;
             // 클리어 비율 (아직 사용안됨)
-            public int ClearRate = 0;
+            public float ClearRate = 0;
         }
 
     }
